* added another refactor to make tests look cleaner * refactored cli to see if it can use run too
