+++ .npmignore
@@ -7,3 +7,4 @@ __tests__
 .editorconfig
 .eslintrc.js
 jest.*
+__mocks__
