+++ __mocks__/argv.js
@@ -1,0 +1,1 @@
+module.exports = (inputs) => (process.argv = inputs);
