+++ __tests__/cli-index.js
@@ -1,3 +1,5 @@
-test('first test', () => {
-  expect(true).toBeTruthy();
+const argv = require('../__mocks__/argv');
+
+test('test argv mock', () => {
+  expect(argv(['--arg1', 'value1'])).toEqual(['--arg1', 'value1']);
 });
