+++ jest.teardown.js
@@ -4,10 +4,6 @@ const { TEST_JSON_DIR } = require('./jest.config');
 
 module.exports = async () => {
   fs.readdir(TEST_JSON_DIR, (_, files) =>
-    files.forEach((file) =>
-      fs.unlink(path.resolve(TEST_JSON_DIR, file), () =>
-        fs.rmdir(TEST_JSON_DIR, () => console.log(`test files deleted`))
-      )
-    )
+    files.forEach((file) => fs.unlink(path.resolve(TEST_JSON_DIR, file), () => fs.rmdir(TEST_JSON_DIR, () => () => _)))
   );
 };
